deep_research/requirements.txt
# Deep Research - LangGraph Multi-Agent System
# Python 3.10+

# Core Framework
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# HTTP & Async
httpx>=0.27.0
aiohttp>=3.9.0
requests>=2.31.0

# Utilities
tenacity>=8.2.0  # Retry logic
structlog>=24.1.0  # Structured logging
rich>=13.7.0  # Console output

# MCP Protocol (if available)
# mcp>=0.1.0

# Testing (optional)
pytest>=8.0.0
pytest-asyncio>=0.23.0
